# VehicleService/api-docs/openapi.json
{
     "openapi": "3.0.0",
     "info": {
          "title": "API Vehicle",
          "version": "v1"
     },
     "servers": [
          {
               "url": "http://localhost:5003"
          }
     ],
     "paths": {
...
          "/vehicle/{vehicleId}": {
               "get": {
                    "description": "presents a vehicle",
                    "tags": ["Vehicle"],
                    "parameters": [
                         {
                              "name": "vehicleId",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "number",
                                   "example": 1
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "A given vehicle",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "array",
                                             "items": {
                                                  "$ref": "#/components/schemas/objVehicleResponse"
                                             }
                                        }
                                   }
                              }
                         }
                    }
               },
               "put": {
                    "description": "Updates the vehicle information",
                    "tags": ["Vehicle"],
                    "parameters": [
                         {
                              "name": "vehicleId",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "number",
                                   "example": 1
                              }
                         }
                    ],
                    "requestBody": {
                         "content": {
                              "application/json": {
                                   "schema": {
                                        "$ref": "#/components/schemas/objVehicleUpdateRequest"
                                   }
                              }
                         }
                    },
                    "responses": {
                         "201": {
                              "description": "Message indication if vehicle was updated",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type":"object",
                                             "properties":{
                                                  "message": {
                                                       "type": "string"
                                                  }
                                             }
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
     },
     "components": {
          "schemas": {
...
               "objVehicleUpdateRequest": {
                    "type": "object",
                    "properties": {
                         "description": {
                              "type": "string"
                         },
                         "vehicleTypeId": {
                              "type": "number"
                         },
                         "available": {
                              "type": "boolean"
                         }
                    }
               },
               "objVehicleResponse": {
                    "type": "object",
                    "properties": {
                         "vehicleId": {
                              "type": "number"
                         },
                         "description": {
                              "type": "string"
                         },
                         "vehicleTypeId": {
                              "type": "number"
                         },
                         "available": {
                              "type": "boolean"
                         },
                         "batery": {
                              "type": "number"
                         }
                    }
               },
...
          }
     },
...
}